Move 'gce' package inside the 'cloudproviders' one
