+++ pkg/util/cloudproviders/gce/diagnosis.go
