+++ pkg/util/cloudproviders/gce/test/gce_metadata.json
